+++ R/add_noise.R
@@ -2,7 +2,7 @@
 #'
 #' \code{add_noise} adds synthetic noise to annotations in extended selection tables
 #' @inheritParams template_params
-#' @param X Object of class 'extended_selection_table' (created by the function \code{\link[warbleR]{selection_table}} from the warbleR package), generated 'by element', with the test sound files' annotations (which in baRulho is typically the output of \code{\link{align_test_files}}). Must contain the following columns: 1) "sound.files": name of the .wav files, 2) "selec": unique selection identifier (within a sound file), 3) "start": start time and 4) "end": end time of selections, 5) "bottom.freq": low frequency for bandpass, 6) "top.freq": high frequency for bandpass and 7) "sound.id": ID of sounds (needed to calculate signal to noise ratio internally using \code{\link{signal_to_noise_ratio}}).
+#' @param X Object of class 'extended_selection_table' (created by the function \code{\link[warbleR]{selection_table}} from the warbleR package), generated 'by element' (see 'https://marce10.github.io/warbleR/articles/b_annotation_data_format.html#by-element-vs-by-song-extended-selection-tables'), with the test sound files' annotations (which in baRulho is typically the output of \code{\link{align_test_files}}). Must contain the following columns: 1) "sound.files": name of the .wav files, 2) "selec": unique selection identifier (within a sound file), 3) "start": start time and 4) "end": end time of selections, 5) "bottom.freq": low frequency for bandpass, 6) "top.freq": high frequency for bandpass. If the "sound.id" column is supplied noise is only added to those sounds with a 'sound.id' different from "ambient", "start_marker" or "end_marker".
 #' @param mar numeric vector of length 1. Specifies the margins adjacent to
 #'   the start point of the annotation over which to measure ambient noise.
 #' @param target.snr numeric vector of length 1. Specifies the desired signal-to-noise ratio. Must be lower that the current signal-to-noise ratio. Annotations showing a signal-to-noise ratio higher than 'target.snr' will remain unchanged. Must be supplied.
@@ -11,7 +11,7 @@
 #' @param kind Character vector of length 1 indicating the kind of noise, “white”, “pink”, “power”, "brown", or “red”. Noise is synthesized with a modified version of the function \code{\link[tuneR]{noise}}. Default is "pink" which is similar to background noise in natural environments.
 #' @param alpha Numeric vector of length 1. The power for the power law noise (defaults are 1 for pink and 1.5 for red noise). Only used when \code{kind = "power"}.
 #' @param seed Numeric vector of length 1. Seed for random number generation. Default is 123. If NULL, the seed is not set.
-#' @param ... Additional arguments to be passed internally to \code{\link{signal_to_noise_ratio}}.
+#' @param ... Additional arguments to be passed internally to \code{\link{signal_to_noise_ratio}}. Note that "custom" noise reference (argument 'noise.ref' in \code{\link{signal_to_noise_ratio}}) is currently not supported.
 #' @return Object 'X' in which the wave objects have been modified to match the target signal-to-noise ratio. It also includes an additional column, 'adjusted.snr', with the new signal-to-noise ratio values.
 #' @export
 #' @name add_noise
@@ -66,8 +66,12 @@ add_noise <-
     # report errors
     .report_assertions(check_results)
     
-    # get index number
-    target_rows <- which(!X$sound.id %in% c("ambient", "marker"))
+    # get index number 
+    # exclude "ambient" sounds if "sound.id" column was supplied
+    target_rows <- if (!is.null(X$sound.id)){
+      which(!X$sound.id %in% c("ambient", "start_marker", "end_marker")) } else {
+        seq_len(nrow(X))
+      }
     
     wav_snr_list <-
       warbleR:::.pblapply(